\name{MortalitySmooth-package}

\alias{MortalitySmooth-package}

\alias{MortalitySmooth}

\docType{package}

\title{
  Smoothing Poisson counts with P-splines
}

\description{
  Smoothing one- and two-dimensional Poisson counts with P-splines
  specifically tailored to mortality data. Extra-Poisson variation can
  be accounted as well as mortality forecasting. Collection of mortality
  data and a specific function for selecting those data by country, sex,
  age and years.
}

\details{
\tabular{ll}{
Package: \tab MortalitySmooth\cr
Type: \tab Package\cr
Version: \tab 2.2\cr
Date: \tab 2012-02-05\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Carlo G Camarda

Maintainer: Carlo G Camarda <camarda@demogr.mpg.de>
}

\references{ 
Camarda, C. G. (2012). {MortalitySmooth}: An {R} Package for Smoothing
{P}oisson Counts with {P}-Splines. \emph{Journal of Statistical
Software}. 50, 1-24. \url{http://www.jstatsoft.org/v50/i01/}.
  
Eilers P. H. C. and B. D. Marx (1996). Flexible Smoothing with {B}-splines and
Penalties. \emph{Statistical Science}. 11, 89-121.

Currie, I. D., M. Durban, and P. H. C. Eilers (2004). Smoothing and forecasting mortality
rates. \emph{Statistical Modelling}. 4, 279-298.

Currie, I. D., M. Durban, and P. H. C. Eilers (2006). Generalized linear
array models with applications to multidimentional
smoothing. \emph{Journal of the Royal Statistical Society. Series
  B}. 68, 259-280.
}

\keyword{ package }

%\examples{
%~~ simple examples of the most important functions ~~
%}
