\name{MortalitySmooth-package}

\alias{MortalitySmooth-package}

\alias{MortalitySmooth}

\docType{package}

\title{
  Smoothing Poisson counts with P-splines
}

\description{
  Smoothing one- and two-dimensional Poisson counts with P-splines
  specifically tailored to mortality data. Extra-Poisson variation can
  be accounted. Collection of mortality data and a specific function for
  selecting those data by country, sex, age and years. 
}

\details{
\tabular{ll}{
Package: \tab MortalitySmooth\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2009-09-17\cr
License: \tab What license is it under?\cr
LazyLoad: \tab yes\cr
}

}
\author{
Carlo G Camarda

Maintainer: Carlo G Camarda <camarda@demogr.mpg.de>
}

\references{ 
Eilers and Marx (1996). Flexible Smoothing with B-splines and
Penalties. Statistical Science. Vol. 11. 89-121.

Currie, Durban and Eilers (2004). Smoothing and forecasting mortality
rates. Statistical Modelling. Vol. 4. 279-298.

Currie, Durban and Eilers (2006). Generalized linear array models with
applications to multidimensional smoothing. JRSS. Vol. 68. 259-280. 
}

\keyword{ package }

%\examples{
%~~ simple examples of the most important functions ~~
%}
